Hand Realignment
Leif, is this what you're looking for? I didn't want to keep going if I
didn't quite have it right.  I think I'm getting the divs aligned
correctly, but I haven't been backspacing to the beginning of each line,
so there may be hidden tabs in the whitespace. Is that a problem?